// src/types/types.ts
export type AdminDataType = {
  userId: null | string;
  company: null | string;
  name: null | string;
  addinional: null | string;
  street: null | string;
  postal_code: null | string;
  country: null | string;
  iban: null | string;
  bic: null | string;
  bank_name: null | string;
  fax: null | string;
  email: null | string;
  birthday: null | string;
  homepage: null | string;
};

export type ModalData = {
  modal1: boolean;
  modal2: boolean;
  modal3: boolean;
};